Rename service package, classes and variables names
