+++ aplicacao-grafica-samu/src/main/java/br/ufrrj/samu/controllers/HomeController.java
@@ -2,24 +2,24 @@ package br.ufrrj.samu.controllers;
 
 import br.ufrrj.samu.entities.Student;
 import br.ufrrj.samu.entities.Subject;
-import br.ufrrj.samu.services.StudentService;
-import br.ufrrj.samu.services.SubjectService;
+import br.ufrrj.samu.repositories.StudentRepository;
+import br.ufrrj.samu.repositories.SubjectRepository;
 
 import java.util.List;
 import java.util.Optional;
 
 public class HomeController {
 
-    private SubjectService subjectService;
-    private StudentService studentService;
+    private SubjectRepository subjectRepository;
+    private StudentRepository studentRepository;
 
-    public HomeController(SubjectService subjectService, StudentService studentService) {
-        this.subjectService = subjectService;
-        this.studentService = studentService;
+    public HomeController(SubjectRepository subjectRepository, StudentRepository studentRepository) {
+        this.subjectRepository = subjectRepository;
+        this.studentRepository = studentRepository;
     }
 
     public Student getStudent(String username) {
-        Optional<Student> studentOptional = studentService.findStudentByUsername(username);
+        Optional<Student> studentOptional = studentRepository.findStudentByUsername(username);
         //Isso nunca vai acontecer, mas tem que seguir as boas práticas né?
         if (studentOptional.isEmpty()) {
             return new Student("João Gameplays e Tutoriais", "QWERTY", "João", "Minha rua, número da minha casa", List.of(), "Curso", "0.1");
@@ -29,7 +29,7 @@ public class HomeController {
     }
 
     public List<Subject> getStudentSubjects(String username) {
-        Optional<Student> studentOptional = studentService.findStudentByUsername(username);
+        Optional<Student> studentOptional = studentRepository.findStudentByUsername(username);
         //Isso nunca vai acontecer, mas tem que seguir as boas práticas né?
         if (studentOptional.isEmpty()) {
             return List.of();
