+++ aplicacao-grafica-samu/src/main/java/br/ufrrj/samu/entities/Subject.java
@@ -1,11 +1,9 @@
 package br.ufrrj.samu.entities;
 
-import br.ufrrj.samu.services.SubjectService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 public class Subject {
